+++ tests/mixin_tests.py
@@ -22,20 +22,18 @@ class MeasurementTestCase(base.AsyncServerTestCase):
         self.assertEqual(measurement.name, 'my-service')
         self.assertEqual(measurement.tags['status_code'], '200')
         self.assertEqual(measurement.tags['method'], 'GET')
-        self.assertEqual(
-            measurement.tags['handler'], 'tests.base.RequestHandler')
+        self.assertEqual(measurement.tags['handler'],
+                         'tests.base.RequestHandler')
         self.assertEqual(measurement.tags['endpoint'], '/')
         self.assertEqual(measurement.tags['hostname'], socket.gethostname())
+        self.assertEqual(measurement.tags['content_type'], 'application/json')
+
         self.assertEqual(measurement.fields['content_length'], 16)
         self.assertGreater(float(measurement.fields['duration']), 0.001)
         self.assertLess(float(measurement.fields['duration']), 0.1)
-        self.assertEqual(measurement.tags['accept'], 'application/json')
-        self.assertEqual(measurement.tags['content_type'], 'application/json')
 
-        nanos_since_epoch = int(measurement.timestamp)
-        then = nanos_since_epoch / 1000000000
-        self.assertGreaterEqual(then, int(start_time))
-        self.assertLessEqual(then, time.time())
+        self.assertGreaterEqual(measurement.timestamp, int(start_time))
+        self.assertLessEqual(measurement.timestamp, time.time())
 
     def test_measurement_with_named_endpoint(self):
         start_time = time.time()
@@ -50,21 +48,17 @@ class MeasurementTestCase(base.AsyncServerTestCase):
         self.assertEqual(measurement.tags['endpoint'], '/named')
         self.assertEqual(measurement.tags['content_type'], 'application/json')
         self.assertEqual(
-            measurement.tags['correlation_id'],
-            base.NamedRequestHandler.correlation_id)
-        self.assertEqual(
             measurement.tags['handler'], 'tests.base.NamedRequestHandler')
         self.assertEqual(measurement.tags['hostname'], socket.gethostname())
         self.assertEqual(measurement.fields['content_length'], 16)
-        self.assertNotIn('accept', measurement.tags)
+        self.assertEqual(
+            measurement.fields['correlation_id'],
+            base.NamedRequestHandler.correlation_id)
 
         self.assertGreater(float(measurement.fields['duration']), 0.001)
         self.assertLess(float(measurement.fields['duration']), 0.1)
-
-        nanos_since_epoch = int(measurement.timestamp)
-        then = nanos_since_epoch / 1000000000
-        self.assertGreaterEqual(then, int(start_time))
-        self.assertLessEqual(then, time.time())
+        self.assertGreaterEqual(measurement.timestamp, int(start_time))
+        self.assertLessEqual(measurement.timestamp, time.time())
 
     def test_measurement_with_param_endpoint(self):
         result = self.fetch('/param/100')