+++ docs/history.rst
@@ -3,6 +3,16 @@
 Release History
 ===============
 
+`1.0.4`_ (14 Sep 2016)
+----------------------
+- Rework how the line protocol is marshalled and support the various data types.
+- Remove the accept tag
+- Strip down content_type to the ``type/subtype`` format only
+- Make ``correlation_id`` a field value and not tag
+- Change the precision to second precision, per the InfluxDB docs (use the most
+    coarse precision for better compression)
+
+
 `1.0.3`_ (13 Sep 2016)
 ----------------------
 - Add a response ``content_length`` field, an ``accept`` tag (if set in request
@@ -20,8 +30,9 @@ Release History
 ----------------------
 - Initial release
 
+.. _1.0.4: https://github.com/sprockets/sprockets-influxdb/compare/1.0.3...1.0.4
 .. _1.0.3: https://github.com/sprockets/sprockets-influxdb/compare/1.0.2...1.0.3
 .. _1.0.2: https://github.com/sprockets/sprockets-influxdb/compare/1.0.1...1.0.2
 .. _1.0.1: https://github.com/sprockets/sprockets-influxdb/compare/1.0.0...1.0.1
 .. _1.0.0: https://github.com/sprockets/sprockets-influxdb/compare/0.0.0...1.0.0
-.. _Next Release: https://github.com/sprockets/sprockets-influxdb/compare/1.0.3...master
+.. _Next Release: https://github.com/sprockets/sprockets-influxdb/compare/1.0.4...master